add sweet alert on delete

+++ src/app/pages/statistic/statistic.component.ts
@@ -24,11 +24,42 @@ export class StatisticComponent implements OnInit {
   }
 
   //delete record method
-  deleteRecord(id:any){     
-       this.crudservice.deleteUserRecord(id).subscribe(data => {
-           this.getUserData = this.getUserData.filter((dr: { id: any; }) => dr.id !== id);
-           this.toasterservice.showError("Record Deleted Successfully !!", "Delete Record");
-       });
+  // deleteRecord(id:any){     
+  //      this.crudservice.deleteUserRecord(id).subscribe(data => {
+  //          this.getUserData = this.getUserData.filter((dr: { id: any; }) => dr.id !== id);
+  //          this.toasterservice.showError("Record Deleted Successfully !!", "Delete Record");
+  //      });
+  // }
+
+ //delete record method using sweet alert box
+  deleteRecord(id:any){
+    Swal.fire({
+      title: 'Are you sure want to remove?',
+      text: 'You will not be able to recover this file!',
+      icon: 'warning',
+      showCancelButton: true,
+      confirmButtonText: 'Yes, delete it!',
+      cancelButtonText: 'No, keep it'
+    }).then((result) => {
+      if (result.value) {
+        Swal.fire(
+          'Deleted!',
+          'Your imaginary file has been deleted.',
+          'success'
+        )
+        //delete based on
+        this.crudservice.deleteUserRecord(id).subscribe(data => {
+          this.getUserData = this.getUserData.filter((dr: { id: any; }) => dr.id !== id);
+          this.toasterservice.showError("Record Deleted Successfully !!", "Delete Record");
+      });
+      } else if (result.dismiss === Swal.DismissReason.cancel) {
+        Swal.fire(
+          'Cancelled',
+          'Your imaginary file is safe :)',
+          'error'
+        )
+      }
+    })
   }
 
 }
